requirements.txt
lxml
httpx
PySide6
pikpakapi
cssselect
python-vlc